Debug.NoTrace

Description

This module introduce functions that have identical types with functions in the Debug.Trace module.

You might write some programs like this:

import Debug.Trace

fib 0 = 1
fib 1 = 1
fib n = ("fib " ++ show n) `trace` fib (n - 1) + fib (n - 2)

And after you finish the debugging process, just change the line

import Debug.Trace

into

import Debug.NoTrace

Then all the tracing functions are silently removed.

Documentation

trace :: String -> a -> a

traceId :: String -> String

traceShow :: Show a => a -> b -> b

traceShowId :: Show a => a -> a

traceStack :: String -> a -> a

traceIO :: String -> IO ()

traceM :: Monad m => String -> m ()

traceShowM :: (Show a, Monad m) => a -> m ()

putTraceMsg :: String -> IO ()

traceEvent :: String -> a -> a

traceEventIO :: String -> IO ()

traceMarker :: String -> a -> a

traceMarkerIO :: String -> IO ()